init: Allow devices to use user-space tools to set ro.serialno
CUrrently, init will always set a ro.serialno value which means
that the kernel is the only option for configuring the serial
number.

The zenfone2 stores the serial number in a file-system that
is mounted during the boot process and provides tools that pull
the factory configuration from there.

To support this, we need to allow init to opt-out of setting
the serial number for us.

Change-Id: I9c307d91bb8d3cb1c881213dd28e4f76e6eae14f
Signed-off-by: Hriday Sharma <hridaysharma42@gmail.com>